fixed code decomposition
